+++ src/main/java/exeptions/ItemAmountUnavailableException.java
@@ -3,7 +3,7 @@ package exeptions;
 public class ItemAmountUnavailableException extends Exception {
 
     int availableAmount;
-    public ItemAmountUnavailableException(String message, int availableAmount) {
+    public ItemAmountUnavailableException(String message, Double availableAmount) {
         super(message);
         this.availableAmount = availableAmount;
     }
